% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_scale.R
\name{min_max_scale}
\alias{min_max_scale}
\title{Min-Max Scaling}
\usage{
min_max_scale(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A normalized numeric vector.
}
\description{
This function normalizes a numeric vector to the range [0,1].
}
\examples{
min_max_scale(c(1,2,3))
}
