% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdcnn_pred.R
\name{sdcnn_pred}
\alias{sdcnn_pred}
\title{Make Predictions Using an SDCNN Model}
\usage{
sdcnn_pred(model, coords, X = NULL, y, venv, num_pred)
}
\arguments{
\item{model}{A trained SDCNN model, must be the output from the function sdcnn_train().}

\item{coords}{A matrix (n x 2) of longitude and latitude values.}

\item{X}{A matrix (n x p) of covariates, or NULL if not used.}

\item{y}{A numeric vector (n x 1) of response values.}

\item{venv}{The name of the Conda virtual environment to use.}

\item{num_pred}{SDCNN has uncertainty quantification. This arguments specifies the number of predictions to generate.}
}
\value{
A matrix of predictions, each column is one prediction, and each row stands for a location.
}
\description{
This function applies a trained SDCNN model to predict the response variable at unobserved locations.
}
\examples{
\dontrun{
predictions <- sdcnn_pred(model, coords, X = NULL, y, venv, num_pred = 5)
}
}
